Day 4/Deploying a RAG System/backend/requirements.txt
fastapi[all]
openai
faiss-cpu
numpy
python-dotenv
pydantic
